moved logic into use hook
