+++ src/state/gameSlice.ts
@@ -31,12 +31,12 @@ const gameStateSlice = createSlice({
     setGameStarted: (state: GameState, action: PayloadAction<boolean>): void => {
       state.gameStarted = action.payload;
     },
-    setPollGetGameCount: (state: GameState, action: PayloadAction<number>): void => {
-      state.pollGetGameCount = action.payload;
+    incrementPollGetGameCount: (state: GameState, action: PayloadAction<number>): void => {
+      state.pollGetGameCount = state.pollGetGameCount + action.payload;
     }
   },
 });
 
-export const { setGameCode, createGame, setHasJoinedGame, setGameStarted, setPollGetGameCount } = gameStateSlice.actions;
+export const { setGameCode, createGame, setHasJoinedGame, setGameStarted, incrementPollGetGameCount } = gameStateSlice.actions;
 
 export const gameSliceReducer = gameStateSlice.reducer;
